# src/swagger/swagger.json
{"openapi":"3.0.1","servers":[{"url":"http://localhost:80","description":"게시판 데모"}],"info":{"title":"ClassTing Homework API","description":"Private으로 지정 된 API는 토큰 검증을 통해 사용되어야 하는 API이며, 해당 API에서는 검증을 구현 하지 않아 프로퍼티로 받아 사용합니다.","version":"1.0.0","license":{"name":"UNLICENSED"}},"paths":{"/school":{"get":{"tags":["school public"],"parameters":[],"responses":{"200":{"description":"등록 된 학교페이지 목록을 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/SuccessResponseArraySchoolDto"},{"$ref":"#/components/schemas/FailedResponse"}]}}}}},"summary":"등록 된 학교페이지 목록을 조회하는 API 입니다","description":"등록 된 학교페이지 목록을 조회하는 API 입니다."},"post":{"tags":["school private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolBody.Create"}}},"required":true},"responses":{"201":{"description":"학교 페이지 생성 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지 생성을 하는 API 입니다","description":"학교 페이지 생성을 하는 API 입니다.\n\n권한이 충족되지 않을 경우 실패를 반환합니다."},"delete":{"tags":["school private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SchoolBody.Delete"}}},"required":true},"responses":{"200":{"description":"학교 페이지 삭제 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지 삭제하는 API 입니다","description":"학교 페이지 삭제하는 API 입니다.\n\n권한이 충족되지 않을 경우 실패를 반환합니다."}},"/school/news":{"get":{"tags":["school public"],"parameters":[{"name":"query","in":"query","schema":{"$ref":"#/components/schemas/NewsQuery.LookUp"},"description":"","required":true}],"responses":{"200":{"description":"발행 된 소식목록을 불러 옵니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseArrayNewsDto"}]}}}}},"summary":"발행 된 학교 소식목록을 조회하는 API 입니다","description":"발행 된 학교 소식목록을 조회하는 API 입니다."},"post":{"tags":["school private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsBody.Create"}}},"required":true},"responses":{"201":{"description":"학교 소식 발행 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지 소식을 발행하는 API 입니다","description":"학교 페이지 소식을 발행하는 API 입니다.\n\n권한이 충족되지 않을 경우 실패를 반환합니다."},"patch":{"tags":["school private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsBody.Update"}}},"required":true},"responses":{"200":{"description":"학교 소식 갱신 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지 소식을 갱신, 업데이트하는 API 입니다","description":"학교 페이지 소식을 갱신, 업데이트하는 API 입니다.\n\n권한이 충족되지 않을 경우 실패를 반환합니다."},"delete":{"tags":["school private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsBody.Delete"}}},"required":true},"responses":{"200":{"description":"학교 소식 삭제 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지 소식을 삭제하는 API 입니다","description":"학교 페이지 소식을 삭제하는 API 입니다.\n\n권한이 충족되지 않을 경우 실패를 반환합니다."}},"/school/news/detail":{"get":{"tags":["school public"],"parameters":[{"name":"query","in":"query","schema":{"$ref":"#/components/schemas/NewsQuery.FindDetail"},"description":"","required":true}],"responses":{"200":{"description":"소식의 상세정보를 찾아 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseNewsDetailDto"}]}}}}},"summary":"학교 페이지 소식의 상세정보를 찾는 API 입니다","description":"학교 페이지 소식의 상세정보를 찾는 API 입니다."}},"/user/feeds":{"get":{"tags":["user private"],"parameters":[{"name":"query","in":"query","schema":{"$ref":"#/components/schemas/UserQuery.LookUpFeeds"},"description":"","required":true}],"responses":{"200":{"description":"내 뉴스피드를 조회 합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseArrayNewsFeedDto"}]}}}}},"summary":"뉴스피드 조회를 위한 API 입니다","description":"뉴스피드 조회를 위한 API 입니다."}},"/user":{"post":{"tags":["user public"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.Create"}}},"required":true},"responses":{"201":{"description":"유저 등록 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"유저 등록을 위한 API 입니다","description":"유저 등록을 위한 API 입니다.\n\n권한 설정은 불가합니다."}},"/user/login":{"post":{"tags":["user private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.Login"}}},"required":true},"responses":{"201":{"description":"성공 시 유저 정보를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseUserDtonull"}]}}}}},"summary":"유저 로그인 API 입니다","description":"유저 로그인 API 입니다."}},"/user/subscribe":{"patch":{"tags":["user private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.Subscribe"}}},"required":true},"responses":{"200":{"description":"학교 페이지 구독 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"학교 페이지를 구독하는 API 입니다","description":"학교 페이지를 구독하는 API 입니다."}},"/user/unsubscribe":{"patch":{"tags":["user private"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBody.UnSubscribe"}}},"required":true},"responses":{"200":{"description":"학교 페이지 구독 취소 결과를 반환합니다.","content":{"application/json":{"schema":{"oneOf":[{"$ref":"#/components/schemas/FailedResponse"},{"$ref":"#/components/schemas/SuccessResponseboolean"}]}}}}},"summary":"구독 한 학교 페이지를 삭제하는 API 입니다","description":"구독 한 학교 페이지를 삭제하는 API 입니다."}}},"components":{"schemas":{"SuccessResponseArraySchoolDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/SchoolDto"}},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"SchoolDto":{"type":"object","properties":{"region":{"type":"string","maxLength":10},"name":{"type":"string","maxLength":20}},"nullable":false,"required":["region","name"]},"FailedResponse":{"type":"object","properties":{"status":{"type":"number"},"message":{"type":"string"}},"nullable":false,"required":["status","message"]},"SchoolBody.Create":{"type":"object","properties":{"school_name":{"type":"string","maxLength":20},"region":{"type":"string","maxLength":10},"register_email":{"type":"string","format":"email"}},"nullable":false,"required":["school_name","region","register_email"]},"SuccessResponseboolean":{"type":"object","properties":{"data":{"type":"boolean"},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"SchoolBody.Delete":{"type":"object","properties":{"register_email":{"type":"string","format":"email"},"school_name":{"type":"string","maxLength":20}},"nullable":false,"required":["register_email","school_name"]},"NewsQuery.LookUp":{"type":"object","properties":{"desc":{"type":"string"},"school_name":{"type":"string","maxLength":20}},"nullable":false},"SuccessResponseArrayNewsDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/NewsDto"}},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"NewsDto":{"type":"object","properties":{"title":{"type":"string"},"detailId":{"type":"string","maxLength":32},"createdAt":{"type":"string","format":"date-time"}},"nullable":false,"required":["title","detailId","createdAt"]},"NewsQuery.FindDetail":{"type":"object","properties":{"detailId":{"type":"string"}},"nullable":false,"required":["detailId"]},"SuccessResponseNewsDetailDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/NewsDetailDto"},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"NewsDetailDto":{"type":"object","properties":{"contents":{"type":"string"},"views":{"type":"number"},"updatedAt":{"type":"string","format":"date-time"}},"nullable":false,"required":["contents","views","updatedAt"]},"NewsBody.Create":{"type":"object","properties":{"school_name":{"type":"string","maxLength":20},"title":{"type":"string","maxLength":50},"contents":{"type":"string"},"register_email":{"type":"string","format":"email"}},"nullable":false,"required":["school_name","title","contents","register_email"]},"NewsBody.Update":{"type":"object","properties":{"newsId":{"type":"string"},"contents":{"type":"string"},"title":{"type":"string","maxLength":50},"register_email":{"type":"string","format":"email"}},"nullable":false,"required":["newsId","register_email"]},"NewsBody.Delete":{"type":"object","properties":{"newsId":{"type":"string"},"register_email":{"type":"string","format":"email"}},"nullable":false,"required":["newsId","register_email"]},"UserQuery.LookUpFeeds":{"type":"object","properties":{"email":{"type":"string","format":"email"},"desc":{"type":"string"}},"nullable":false,"required":["email"]},"SuccessResponseArrayNewsFeedDto":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/NewsFeedDto"}},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"NewsFeedDto":{"type":"object","properties":{"school_name":{"type":"string","maxLength":20},"news":{"$ref":"#/components/schemas/NewsDto"}},"nullable":false,"required":["school_name","news"]},"UserBody.Create":{"type":"object","properties":{"pass":{"type":"string","maxLength":34},"email":{"type":"string","format":"email"}},"nullable":false,"required":["pass","email"]},"UserBody.Login":{"type":"object","properties":{"pass":{"type":"string","maxLength":34},"email":{"type":"string","format":"email"}},"nullable":false,"required":["pass","email"]},"SuccessResponseUserDtonull":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserDto.Nullable"},"status":{"type":"number"}},"nullable":false,"required":["data","status"]},"UserDto.Nullable":{"type":"object","properties":{"email":{"type":"string","format":"email"},"authority":{"$ref":"#/components/schemas/Authority"},"subschool":{"type":"array","items":{"type":"string"}}},"nullable":true,"required":["email","authority","subschool"]},"Authority":{"type":"string","enum":["student","admin"]},"UserBody.Subscribe":{"type":"object","properties":{"school_name":{"type":"string","maxLength":20},"region":{"type":"string","maxLength":10},"email":{"type":"string","format":"email"}},"nullable":false,"required":["school_name","region","email"]},"UserBody.UnSubscribe":{"type":"object","properties":{"school_name":{"type":"string","maxLength":20},"email":{"type":"string","format":"email"}},"nullable":false,"required":["school_name","email"]}}}}